Modify the implementation of stack-check.
When the LA backend implements this function, it stores values in memory without
first lowering the stack. The memory detection tool will check whether the
address of the data to be accessed is beyond the range of the stack, so the LA
architecture check will report an error.
So the modified code is the first stack down after the legitimacy check.

gcc/ChangeLog:

	* config/loongarch/loongarch.cc (loongarch_emit_probe_stack_range):
	Modify the implementation of stack-check.
	(loongarch_expand_prologue): Modify the implementation of stack-check.